📝 Cleanup and rename cookbook pages
