+++ apps/cookbook/src/app/examples/link-example/link-example.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'cookbook-link-example',
-  templateUrl: './link-example.component.html',
+  template: '',
   styleUrls: ['../_examples.shared.scss'],
 })
 export class LinkExampleComponent {}
